Move generic geometry things into geom module
